New "Item" entity

+++ src/characters.h
@@ -27,7 +27,7 @@
 // Characters
 Entity obj_character[MAX_CHR];
 Sprite *spr_chr[MAX_CHR];
-u8 active_character; // Which character is the active one
+u16 active_character; // Which character is the active one
 bool movement_active; // Can you move ?
 
 // Faces
@@ -38,16 +38,16 @@ Sprite *spr_face[MAX_FACE];
 #define MAX_COLLISIONS 30 // If we collide more than X times, move the character anyway (CHANGE IT!)
 u8 num_colls;
 
-void init_character(u8 nchar); // Initialize a character
-void release_character(u8 nchar); // Release a character from memory
-void init_face(u8 nface); // Initialize a face
-void release_face(u8 nface); // Release a face from memory
-void update_character(u8 nchar); // Update a character based on every parameter
-void show_character(u8 nchar, bool show); // Show or hide a character
-void anim_character(u8 nchar, u8 newanimation); // Change a character's animation
-void look_left(u8 nchar, bool left); // Make a character look to the left (or right)
-void move_character(u8 nchar, s16 x, s16 y); // Move a character to a new position
-void move_character_instant(u8 nchar, s16 x, s16 y); // Move a character to a new position (instantly)
+void init_character(u16 nchar); // Initialize a character
+void release_character(u16 nchar); // Release a character from memory
+void init_face(u16 nface); // Initialize a face
+void release_face(u16 nface); // Release a face from memory
+void update_character(u16 nchar); // Update a character based on every parameter
+void show_character(u16 nchar, bool show); // Show or hide a character
+void anim_character(u16 nchar, u8 newanimation); // Change a character's animation
+void look_left(u16 nchar, bool left); // Make a character look to the left (or right)
+void move_character(u16 nchar, s16 x, s16 y); // Move a character to a new position
+void move_character_instant(u16 nchar, s16 x, s16 y); // Move a character to a new position (instantly)
 void update_sprites_depth(void); // Update characters and enemies depth
 u16 detect_char_collision(u16 nchar, u16 x, u8 y); // Detect collisons between a character in every enemy, given some new coordinates
